Adds continious playback support
Allows to start watching items from the point on the list to the end
of the list, works by detecting the end of the process, so your media
player must end when file is endet or you have to quit it manually.

Basically it playes the search result like it's a playlist.